Use Redash default (and often lower) process count
Prior to this commit, we were using the processor count to set the
number of processes in the web service. That was causing RAM and/or CPU
to be maxed out. This commit uses the Redash defaults which are defined
here https://github.com/getredash/redash/blob/46ea3b1f0b263ba310de7c2c17e4a32d3ea4c343/bin/docker-entrypoint#L35